Fix Custody.create() revert: Channel struct uses uint64, not uint256
CUSTODY_ABI declared challenge/nonce as uint256 but the Solidity
contract (Types.sol) uses uint64. This produced a wrong function
selector (0xa1e5f023 vs 0x4a7e7798), causing every create() call
to revert.

+++ app/web/src/config/contracts.ts
@@ -173,8 +173,8 @@ export const CUSTODY_ABI = [
         components: [
           { name: 'participants', type: 'address[]' },
           { name: 'adjudicator', type: 'address' },
-          { name: 'challenge', type: 'uint256' },
-          { name: 'nonce', type: 'uint256' },
+          { name: 'challenge', type: 'uint64' },
+          { name: 'nonce', type: 'uint64' },
         ],
       },
       {
